requirements.txt
geopy==2.4.1
piexif==1.1.3
Pillow==10.4.0
pandas~=2.2.2
mediameta
ffmpeg-python
pyinstaller
pillow-heif
